+++ server/lib/server/src/grpc/mod.rs
@@ -16,6 +16,7 @@
 
 pub mod auth;
 mod buffer;
+mod command;
 mod debug;
 mod editor;
 mod extension;
@@ -29,9 +30,9 @@ mod state;
 mod syntax;
 
 pub use {
-    auth::AuthInterceptor, buffer::BufferServiceImpl, debug::DebugServiceImpl,
-    editor::EditorServiceImpl, extension::ExtensionServiceImpl, input::InputServiceImpl,
-    module::ModuleServiceImpl, notification::NotificationServiceImpl,
+    auth::AuthInterceptor, buffer::BufferServiceImpl, command::CommandServiceImpl,
+    debug::DebugServiceImpl, editor::EditorServiceImpl, extension::ExtensionServiceImpl,
+    input::InputServiceImpl, module::ModuleServiceImpl, notification::NotificationServiceImpl,
     presence::PresenceServiceImpl, server_service::ServerServiceImpl, state::StateServiceImpl,
     syntax::SyntaxServiceImpl,
 };