+++ shared/protocol/build.rs
@@ -27,6 +27,7 @@ fn compile_protos() {
                 "proto/reovim/v2/presence.proto",
                 "proto/reovim/v2/debug.proto",
                 "proto/reovim/v2/extension.proto",
+                "proto/reovim/v2/command.proto",
             ],
             &["proto"],
         )